rename package "budgetfreakapplication" to "application"
