+++ server/application/src/main/java/de/budgetfreak/application/BudgetfreakApplication.java
@@ -1,0 +1,27 @@
+package de.budgetfreak.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+@SpringBootApplication
+public class BudgetfreakApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BudgetfreakApplication.class, args);
+    }
+
+    //TODO change that to custom mapping (not globaly!)
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurerAdapter() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**");
+            }
+        };
+    }
+}
