+++ server/application/src/main/java/de/budgetfreak/application/account/domain/AccountRepository.java
@@ -1,0 +1,13 @@
+package de.budgetfreak.application.account.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+/**
+ * Repository for {@link Account}-entities.
+ */
+public interface AccountRepository extends JpaRepository<Account, Long> {
+
+    List<Account> findByUserId(long userId);
+}
