+++ server/application/src/main/java/de/budgetfreak/application/user/UserResource.java
@@ -1,0 +1,25 @@
+package de.budgetfreak.application.user;
+
+import org.springframework.hateoas.ResourceSupport;
+
+public class UserResource extends ResourceSupport {
+
+    private String name;
+    private String currency;
+
+    public UserResource() {
+    }
+
+    public UserResource(String name, String currency) {
+        this.name = name;
+        this.currency = currency;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getCurrency() {
+        return currency;
+    }
+}
