+++ server/application/src/main/java/de/budgetfreak/application/user/domain/UserRepository.java
@@ -1,0 +1,10 @@
+package de.budgetfreak.application.user.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    User findByName(String name);
+}
