--- server/application/src/main/java/de/budgetfreak/budgetfreakapplication/user/UserResource.java
@@ -1,25 +1,0 @@
-package de.budgetfreak.budgetfreakapplication.user;
-
-import org.springframework.hateoas.ResourceSupport;
-
-public class UserResource extends ResourceSupport {
-
-    private String name;
-    private String currency;
-
-    public UserResource() {
-    }
-
-    public UserResource(String name, String currency) {
-        this.name = name;
-        this.currency = currency;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getCurrency() {
-        return currency;
-    }
-}
