--- server/application/src/main/java/de/budgetfreak/budgetfreakapplication/user/domain/UserRepository.java
@@ -1,10 +1,0 @@
-package de.budgetfreak.budgetfreakapplication.user.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    User findByName(String name);
-}
