+++ server/application/src/test/resources/de/budgetfreak/application/user/users.sql
